+++ src/main/java/nextstep/ladder/domain/ladder/LineFactory.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain.line;
+package nextstep.ladder.domain.ladder;
 
 public interface LineFactory {
