+++ src/main/java/nextstep/ladder/domain/ladder/RandomLineFactory.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain.line;
+package nextstep.ladder.domain.ladder;
 
 import nextstep.ladder.util.StreamUtils;
