+++ src/test/java/nextstep/ladder/domain/ladder/LineTest.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain.line;
+package nextstep.ladder.domain.ladder;
 
 import nextstep.ladder.dto.Connections;
 import org.junit.jupiter.api.DisplayName;
